services/exams-service/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.5
alembic==1.13.0

# Google Cloud
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.1
google-auth==2.23.0

# Utilities
httpx==0.25.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pillow==10.1.0

# Health checks and monitoring
healthcheck==1.3.3
